api-gateway/requirements.txt
confluent-kafka==1.8.2
Flask==2.3.3
Flask-JWT-Extended==4.7.1
marshmallow==3.19.0
requests==2.28.2
python-dotenv==1.0.0
grpcio==1.53.0
grpcio-tools==1.53.0
protobuf==4.22.3